requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
click==8.3.1
fastapi==0.135.3
frozenlist==1.8.0
h11==0.16.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.2
slack-sdk==3.41.0
starlette==1.0.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.42.0
yarl==1.23.0
